paymenr/pyproject.toml
[tool.poetry]
name = "paymenr"
version = "0.1.0"
description = ""
authors = ["Muhammad Hamid <61172566+hamidkhans651@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
stripe = "^9.11.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.3"
httpx = "^0.27.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
